Added images and structure to project sections. I also applied external CSS styling
Utilised the img> element to include photos in each project part while giving accessible alt text. Each project is set up as a section containing a link to the project or case study, a title (in h3> format), a picture, and a summary or description (in p> format). Filled in the fictitious picture sources and project component descriptions. To apply uniform styling, I built an external CSS file and connected it to the HTML text. Created a unified and aesthetically pleasing layout for the project by styling the sections, titles, photos, paragraphs, and links. To improve the presentation of the project sections, I changed the placement, margins, padding, and typefaces.